Explicit plugins (#81)
* Add explicit-load versions of plugins

The existing aws plugins install their metadata at package load
time. They aren't meant to be loaded in test/development environments,
but there is no good way to only load them in production since you
can't conditionally load a package.

- move the plugins under a new /awsplugins path and change init() to
  Init() so user must explicitly call a method to load the plugin.
- leave the old plugins in place, delegating to the new plugins. i
  also left the "Origin" constants since those are part of the
  packages' public APIs.

* Add myself to contributors

* Add README.md note about new plugins
